Implement route existence check and reorder middleware in index.ts: - Added checkRouteExists middleware to verify API routes before authentication.
